refactored xml utilities to be shared by original scripts and pipe steps
