+++ pipelines/step_syn_parser.py
@@ -70,6 +70,7 @@ class CCGSynParser(TransformerMixin):
         parse_command = shlex.split(self.ccg_parse.format(input_file, output_file))
         try:
             completed_process = subprocess.run(parse_command, check=True)
+            # transccg_root is the root element, not the entire document
             transccg_root, encoding = translate_candc_tree(output_file, self.log_file)
             parse_data = ParseData(parse_result=transccg_root, 
                                    parse_encode=encoding,
